- Updating ArcadiaEngine.h for use with the Winsock Library - Updating the font manager to get rid of a warning - Updating UI elements to handle 'create and destroy' more adequately - Updating the master server list system to to latest engine code - Added handling of the UI when servers or clients connect - Added the ability to get the IP of a socket

+++ MasterServerList/FontManager.h
@@ -126,7 +126,7 @@ inline unsigned int Font::GetCharacterCountBeforePassingWidth(const char* text, 
 		auto new_return_value = new_string.find_last_of(" ");
 		if ((new_return_value != std::string::npos) && (new_return_value + 1 < new_string.size()))
 		{
-			character_count = new_return_value + 1;
+			character_count = (unsigned int)(new_return_value + 1);
 		}
 	}
 	return character_count;
